A+B (Trial Problem)
You are given two integers a and b. Print a+b.

Input :
The first line contains an integer t (1≤t≤104) — the number of test cases in the input. Then t test cases follow.

Each test case is given as a line of two integers a and b (−1000≤a,b≤1000).

Output :
Print t integers — the required numbers a+b.

Example :
Input :
4
1 5
314 15
-99 99
123 987

Output :
6
329
0
1110